+++ blockstream/index.js
@@ -1,0 +1,3 @@
+const getBlockstreamVout = require('./get_blockstream_vout');
+
+module.exports = {getBlockstreamVout};