+++ index.js
@@ -1,4 +1,5 @@
 const {getAccountingReport} = require('./report');
+const {getChainTransactions} = require('./records');
 const {rateProviders} = require('./fiat');
 
-module.exports = {getAccountingReport, rateProviders};
+module.exports = {getAccountingReport, getChainTransactions, rateProviders};