+++ records/index.js
@@ -1,4 +1,5 @@
 const getAllInvoices = require('./get_all_invoices');
 const getAllPayments = require('./get_all_payments');
+const getChainTransactions = require('./get_chain_transactions');
 
-module.exports = {getAllInvoices, getAllPayments};
+module.exports = {getAllInvoices, getAllPayments, getChainTransactions};